fix: Convert Checkout page to Checkout modal
