+++ src/components/Checkout/components/AddressModal/AddressModal.css
@@ -1,29 +1,21 @@
-.address-modal-container {
+/* .address-modal-container {
   position: fixed;
   inset: 0;
-  z-index: 6;
+  z-index: 1000;
   background-color: rgba(0, 0, 0, 0.2);
   margin: auto;
-}
+  display: flex;
+  flex-direction: column;
+} */
 
-.address-modal-container .address-input-container {
-  position: absolute;
-  inset: 0;
-  background-color: rgb(255, 255, 255);
-  margin: auto;
-  width: 290px;
-  height: 520px;
+.address-input-container {
+  display: flex;
+  flex-direction: column;
   text-align: center;
   border-radius: 10px;
 }
 
-.address-input-container h1 {
-  margin-top: 25px;
-  color: rgb(33, 33, 33);
-  margin-bottom: 25px;
-}
-
-.address-modal-container .address-input-container input {
+.address-input-container input {
   padding: 8px 16px;
   display: block;
   width: 90%;
@@ -32,42 +24,43 @@
   border-radius: 5px;
 }
 
-.address-modal-container .submit {
+.submit {
   color: white;
   background-color: var(--primary-color);
   font-weight: 600;
   display: inline;
 }
 
-.address-modal-container .submit:hover {
+.submit:hover {
   cursor: pointer;
   background-color: black;
 }
-.address-modal-container .btn-container {
+
+.btn-container {
   display: flex;
   justify-content: space-between;
   padding: 0 20px;
 }
 
-.address-modal-container .btn-container button {
+.btn-container button {
   border: none;
   padding: 8px 16px;
 
   border-radius: 5px;
 }
 
-.address-modal-container .btn-container button:hover {
+.btn-container button:hover {
   cursor: pointer;
   background-color: var(--secondary-background-color-onhover);
 }
 
-.address-modal-container .btn-container button:last-child {
+.btn-container button:last-child {
   flex-grow: 1;
   margin-left: 10px;
 }
 
 @media (min-width: 768px) {
-  .address-modal-container .address-input-container{
+  .address-input-container{
     width: 390px;
   }
