+++ src/components/Checkout/components/BillingSummary/BillingSummary.css
@@ -1,7 +1,6 @@
 .billing-container {
-  border-top: 1px solid rgb(226, 226, 226);
   padding: 15px;
-  margin: 0 15px;
+  margin: 10px 15px 0;
 }
 
 .billing-container h1 {
@@ -12,7 +11,8 @@
   margin-bottom: 10px;
   display: flex;
   justify-content: space-between;
-  color: grey
+  color: grey;
+  font-size: 1rem;
 }
 
 .price-details-container > *:first-child {
@@ -22,6 +22,5 @@
 .price-details-container > *:last-child {
   margin-top :10px;
   font-weight: 700;
-  font-size: 35px;
   color: black
 }
