+++ src/components/Checkout/components/DeliveryAddress/DeliveryAddress.css
@@ -1,7 +1,6 @@
 .delivery-address-container {
   border-top: 1px solid rgb(231, 231, 231);
-  padding: 15px;
-  margin: 15px;
+  padding: 15px 0;
 }
 
 .delivery-address-description {
@@ -18,23 +17,6 @@
 }
 
 .delivery-address-container p {
-  color: grey;
   margin-bottom: 15px;
   font-weight: bolder;
-}
-
-.place-order-btn {
-  border: none;
-  color: white;
-  background-color: var(--primary-color);
-  padding: 12px 36px;
-  border-radius: 10px;
-  font-weight: 600;
-  font-size: 1.1rem;
-  margin-top: 50px;
-}
-
-.place-order-btn:hover {
-  cursor: pointer;
-  background-color: black;
 }
