+++ src/pages/UserProfile/Addresses/Addresses.jsx
@@ -7,7 +7,7 @@ import "./Addresses.css";
 import { RiAddFill } from "react-icons/ri";
 import { useState } from "react";
 import { toast } from "react-hot-toast";
-import { AddressModal } from "../../Checkout/components/AddressModal/AddressModal";
+import { AddressModal } from "../../../components/Checkout/components/AddressModal/AddressModal";
 
 export const Addresses = () => {
   const [loading, setLoading] = useState(false);
